+++ src/lib.rs
@@ -2,3 +2,5 @@ pub mod calculate_mean;
 pub mod goals;
 pub mod saving;
 pub mod subject;
+pub mod grades;
+pub mod errors;